ai-service/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
httpx>=0.27.2
beautifulsoup4>=4.12.3
pdfplumber>=0.11.4
python-docx>=1.1.0
groq>=0.5.0
jinja2>=3.1.4
python-dotenv>=1.0.1
playwright>=1.44.0
hypothesis>=6.112.0
pytest>=8.3.0
pytest-asyncio>=0.24.0
lxml>=5.3.0
python-multipart>=0.0.12
